move 410 Gone response in a util file
